refactor: move theme config files to user module
