+++ docs/dev-notes.md
@@ -11,9 +11,9 @@
 - iOS chrome browser bounces when scrolling to bottom and reverts to top. doesn't happen on other browsers, not even brave
 - Write docs on how to add new themes
 - Figure out how to apply .dark class to all dark themes
-- Make Clerk appearnce variables work with new theme
 - Get fonts working
 - move config themes to user module
+- Create user preferences page that includes theme switcher: https://colorswitchcn.heywinit.me/
 
 ```
     "vercel-build": "if [ \"$VERCEL_ENV\" = \"production\" ]; then drizzle-kit push; elif [ \"$VERCEL_ENV\" = \"preview\" ]; drizzle-kit push --force; fi && next build",
